requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
ConfigArgParse==1.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-cacheops==6.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-redis==4.12.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==8.1.3
flake8==3.9.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
funcy==1.16
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.1.0
idna==2.10
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==2.11.3
locust==1.5.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
psutil==5.8.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pymemcache==3.4.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
pywin32==300
pyzmq==22.0.3
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
